Types to keep track of analytics related to LearningParameters
The analytics consist of the accuracies achieved after every iteration,
a measure for the size of the parameters (big parameters are a sign of
overfitting)
and a measure for the amount of change of the parameters.
This will be used to implement an algorithm which automates the process
of optimising learning parameters.